requirements.txt
Pillow==10.0.0
requests==2.31.0
sphinx==7.2.6
build==1.0.3
twine==4.0.2